Move EventProviderWithBlacklist to seperate file
